Finish modularization
This separated timing from the camera parameters and moved camera control
logic, timing logic, and the like out of the main module.  Now lots of
things are a lot less ugly.

This commit is probably (hopefully) still fully functional.  However,
the GUI is still scattered.  That's a fix for some day when I have a ton
of boredom-filled time.